整理mysql数据类型与CSharp的类型映射，实现分页查询

+++ Model.Common/Entity/Column.cs
@@ -11,13 +11,13 @@ namespace Model.Common.Entity
         public string DatabaseName { get; set; }
         public string TableName { get; set; }
         public string Name { get; set; }
-        public int Order { get; set; }
+        public long Order { get; set; }
         public object Default { get; set; }
         public bool IsNullable { get; set; }
         public string DataType { get; set; }
-        public int? CharacterMaxLength { get; set; }
-        public int? NumbericPrecision { get; set; }
-        public int? DatetimePrecision { get; set; }
+        public long? CharacterMaxLength { get; set; }
+        public long? NumbericPrecision { get; set; }
+        public long? DatetimePrecision { get; set; }
         public bool IsPrimaryKey { get; set; }
         public bool IsAutoIncrease { get; set; }
